[vulkan] Fix warnings: static_cast, remove unused
ghstack-source-id: 465d0ae5517db8a59e69e2edb753d0131c0eb3cf
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42195

+++ aten/src/ATen/native/vulkan/Vulkan.cpp
@@ -1041,7 +1041,8 @@ void copy_buffer_to_image(const VBuffer& buffer, VImage& image) {
     int32_t w;
     int32_t h;
   };
-  const ConstBlock constBlock{image.w(), image.h()};
+  const ConstBlock constBlock{static_cast<int32_t>(image.w()),
+                              static_cast<int32_t>(image.h())};
   VBuffer constBuffer =
       makeUniformConstBuffer(&constBlock, sizeof(constBlock));
 
@@ -1102,7 +1103,8 @@ void copy_image_to_buffer(
     int32_t w;
     int32_t h;
   };
-  const ConstBlock constBlock{image.w(), image.h()};
+  const ConstBlock constBlock{static_cast<int32_t>(image.w()),
+                              static_cast<int32_t>(image.h())};
   VBuffer constBuffer =
       makeUniformConstBuffer(&constBlock, sizeof(constBlock));
